Folk-API/pyproject.toml
[tool.poetry]
name = "folk"
version = "0.1.0"
description = "SaaS multi-tenant para gestion y calificacion de concursos de baile."
authors = ["Codex <codex@example.com>"]
readme = "README.md"
packages = [
    { include = "config" },
    { include = "competition" },
]

[tool.poetry.dependencies]
python = "^3.11"
django = ">=5.0,<6.0"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.10"
celery = "^5.4.0"
redis = "^5.2.1"
django-environ = "^0.11.2"
django-anymail = "^12.0"
django-cors-headers = "^4.6.0"
djangorestframework-simplejwt = "^5.5.1"
pillow = "^12.1.1"

[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"
